TST: Moved test_fourier_derivative to test_derivatives.py
